// packages/core/src/api/Client.ts
export declare namespace ProofStateAPIClient {
  export interface Options {
    environment: core.Supplier<string>;
    /** Specify a custom URL to connect the client to. */
    baseUrl?: core.Supplier<string>;
    username?: core.Supplier<string | undefined>;
    password?: core.Supplier<string | undefined>;
    /** Override the X-ProofState-Sdk-Name header */
    xProofStateSdkName?: core.Supplier<string | undefined>;
    /** Override the X-ProofState-Sdk-Version header */
    xProofStateSdkVersion?: core.Supplier<string | undefined>;
    /** Override the X-ProofState-Public-Key header */
    xProofStatePublicKey?: core.Supplier<string | undefined>;
    /** Additional headers to include in requests. */
    headers?: Record<
      string,
      string | core.Supplier<string | null | undefined> | null | undefined
    >;
  }

  export interface RequestOptions {
    /** The maximum time to wait for a response in seconds. */
    timeoutInSeconds?: number;
    /** The number of times to retry the request. Defaults to 2. */
    maxRetries?: number;
    /** A hook to abort the request. */
    abortSignal?: AbortSignal;
    /** Override the X-ProofState-Sdk-Name header */
    xProofStateSdkName?: string | undefined;
    /** Override the X-ProofState-Sdk-Version header */
    xProofStateSdkVersion?: string | undefined;
    /** Override the X-ProofState-Public-Key header */
    xProofStatePublicKey?: string | undefined;
    /** Additional query string parameters to include in the request. */
    queryParams?: Record<string, unknown>;
    /** Additional headers to include in the request. */
    headers?: Record<
      string,
      string | core.Supplier<string | null | undefined> | null | undefined
    >;
  }
}